put docs of each state inside its docstring.
An object's docstring can be accessed from its constructor or from any of its instances, using the `__doc__` attribute. That looks better, makes it more clear where the docs are and should be defined, and kills two birds with one stone, because it also describes the object of anyone reading the code.